(Task 222) Setting up Wagtail in Django
Since Wagtail is meant to be in the back-end with Django, I modified the Django back-end to include everything that Wagtail will need to run, other than a virtual environment and installing the dependencies through there. There might be a merge conflict, depending on what the rest of the team is working on.