(Task 222) Added myself as a superuser and created base html page
While the base HTML page is not complete yet, at least it is there, and I had to modify Django to allow Wagtail to use the "Wagtail" directory of the project. Other than that, I have added myself as a superuser to be able to see the Wagtail admin page (under "localhost:8000/cms").